moved print_output into helper, removed unnecessary use statements
